Refactoring code
fixes https://github.com/indmind/rin/issues/42 and renamcommand to urban
