+++ src/commands/scss.js
@@ -1,6 +1,7 @@
-import Rin from '../core/rin'
 import sass from 'node-sass'
 import beautify from 'cssbeautify'
+
+import { isEmpty, code as markCode } from '../core/rin'
 
 export default class Scss {
     constructor() {
@@ -18,7 +19,7 @@ export default class Scss {
     async handle(command) {
         const code = command.join(' ')
 
-        if (Rin.isEmpty(code)) return this.VARIABLE.codeEmpty
+        if (isEmpty(code)) return this.VARIABLE.codeEmpty
 
         try {
             const response = sass.renderSync({
@@ -30,7 +31,7 @@ export default class Scss {
                 autosemicolon: true
             })
 
-            return Rin.code('css', result)
+            return markCode('css', result)
         } catch (err) {
             return err.message
         }
